terraform: fix mock provider ImportState race
This commit fixes a bug in the mock provider that caused a data race
during tests. The ImportResourceStateResponse.ImportedResources slice from the mock provider was being written by the mock setter and read by the actual resource instance.

During real provider operation, there is no sharing of data, so the race
condition is not present outside testing with the mock provider.