rpcapi: Stacks.ApplyStackChanges initial implementation
This wires up the stackruntime.ApplyPlan function to the RPC API with a
similar strategy as we used for Stacks.PlanStackChanges wrapping
stackruntime.Plan.

This is not well tested at all and so probably doesn't quite work right.
I'm currently optimizing for getting a reasonable stub of everything into
place so that we can have multiple people work together on getting the
details sorted afterwards.