stackruntime: Persist a component's plantimestamp in the plan
We need to preserve the module runtime's chosen plan timestamp through to
the apply phase so that plantimestamp function calls during the apply
phase will return the same result as during the planning phase.

This also means that we will now no longer emit a "component instance"
planned change object if the planning operation fails so much that there
is no plans.Plan object returned, but that's at least consistent with
the fact that we also don't report any resource instance changes in the
same situation. Event emission is always best-effort when the result
has errors.