stackeval: ApplyPlan's ChangeExec must be in main task
The ChangeExec call creates new promises, one of which the caller is
responsible for resolving by calling the "begin" callback.

To correctly follow the rules of the "promising" package we must resolve
that promise in the same task that created it or explicitly delegate it
to another task. In this case it's fine to just resolve it, but for that
to work correctly we need to call "begin" from the same task as ChangeExec
was called from.