stackeval: InputVariable should respect EvalPhase when deciding value
Previously it was still hard-coding PlanPhase, which was a cheat we did
in earlier work when we were just exploring the overall design of this
package.

It will now pass ApplyPhase when running in the apply phase, although
Main.RootVariableValue doesn't yet have any implementation for retrieving
variable values frozen during planning so root input variables will always
be unknown values during the apply phase for now. We'll fix that in a
future commit.