stackeval: NewForApplying must save the ChangeExecResults object
This was just an oversight in some earlier refactoring. An ApplyPhase
instance of Main must store a pointer to the ChangeExecResults object that
will capture all of the results of the main side-effects of applying the
plan, since we'll need to retrieve those results to evaluate expressions
and other similar downstream results based on those outcomes.