stacks: A very basic stub of applied resource instance change announcements
We'd like to start developing some clients for this part of the RPC API
concurrently with remaining work here in Terraform Core, and so this is
a bare-minimum implementation of emitting "applied change" events for
resource instances just so there's something for client developers to test
against.

There are various things wrong with this, including but not limited to:
 - It always reports that the new state for a resource instance is an
   empty object, rather than including the correct updated object.
 - It only supports "current" resource instance objects, and ignores
   deposed ones.
 - It just uses the resource instance address alone as the key for the
   state description map, which is fine for now when we only have one
   kind of description object anyway but will not be sufficient for a
   real implementation that needs to emit various different kinds of
   object.

We'll need to rework most of this in future commits, but this is hopefully
sufficient to start implementing and testing API clients, at least to some
extend, despite the crudity of the results.