states: ParseDeposedKey function
This is a helper for when we're accepting a deposed key from outside of
Terraform, such as in a plan or state file. It doesn't really "parse"
the string in the usual sense -- the result is always just the input
verbatim if successful -- but it's named that way for consistency with
how we usually name functions like this.