stackstate/statekeys: Representation of stacks state keys
Our model for state for a stack involves a set of objects that is each
identified by an opaque string key. Although those keys are opaque to
callers of Terraform Core, we will actually be using them for some meaning
in Terraform Core itself, since that will avoid redundantly storing the
same information both in the key and in the object associated with the
key.

This therefore aims to encapsulate the generation and parsing of these
keys to help ensure we'll always use them consistently.