Change release order of operations
This PR moves the important TFC upload workflow higher up the release chain, so that it immediately happens after the binaries are published to the Releases Site, and the GitHub release is published. This will allow the team to perform other post-release steps (announcements, etc) without needing to wait for the more time-consuming linux packaging & docker publishing workflows to complete.