core: Minimal working implementation of provider-contributed functions
This establishes all of the wiring necessary for a provider's declared
functions to appear in the hcl.EvalContext when we evaluate expressions
inside a module that depends on a provider that contributes functions.

This does not yet make any attempts to guarantee that the
provider-contributed functions correctly honor the contracts such as
behaving as pure functions. Properly checking this is important because
if a function doesn't uphold Terraform's expectations then it will cause
confusing errors reported downstream, incorrectly blaming other
components for the inconsistency.