stackstate: Initial work on decoding and the state model
This is a sketch of the overall structure of the prior state decoder and
the model type it populates.

Before we can complete this we'll need to slightly rework how the apply
phase emits the raw events that this is consuming, and in particular to
change the raw state representation to be JSON-based to match with how
Terraform Core expects to receive it once reloaded. That will follow in
later commits.