stackruntime: ComponentInstance considers previous state
Now that we have access to a previous state snapshot during the planning
phase, we can pass the relevant parts to the modules runtime when asking
it to plan changes for a component instance.

This is not yet complete enough for full support of consecutive plan and
apply runs that preserve state, but it gets the needed data plumbed in
and we'll continue making this more complete in subsequent commits.