Refactor `(*Cloud).Configure`, and fix TF_WORKSPACE bug
This fixes issue #33976, introduced in https://github.com/hashicorp/terraform/pull/33489,
which broke the intended behavior of specifying the active workspace via the
TF_WORKSPACE variable when using a tag-based workspace mapping.

Now that all the default and fallback value behaviors are cleanly isolated in a
function, this whole flow can be a bit simpler.

- Remove `setConfigurationFields`. Instead, `Configure` can just trade its Cty
  `obj` for a dumb struct and set a couple fields from it. The
  `TF_FORCE_LOCAL_BACKEND` handling can just join the relevant section of
  Configure directly.
- Radically chop down PrepareConfig. It turns out we were violating the interface
  contract, which says PrepareConfig shouldn't trouble itself with the shell
  environment and fallback values... So, don't do that.