stackruntime: Emit some more progress hooks during the apply phase
This is not quite comprehensive yet, but echoes the main component
start/end events we emit during the planning phase, and some of the
resource-instance-related events.

The rpcapi layer is not yet consuming all of these events and so most of
this will just be no-op right now; we'll add the rpcapi translation layer
for these in subsequent commits.