stackruntime: Emit "pending" events for component instance plan/apply
Previously we did have an event for planning but it wasn't in quite the
right place, and so could potentially appear "late" and mess up the
event sequence. Now we'll always emit the pending event immediately before
the planning or applying event, which is perhaps a little redundant but
having these separated means that we might be able to create some space
between them later if we find that we need to do some additional work
after we know about a component instance but before we're reading to
say that we're planning or applying it.