rpcapi(stacks): Additional OpenTelemetry tracing metadata
During PlanStackChanges and ApplyStackChanges we'll mark the trace span
representing the planning or applying of a particular component instance
as either OK or Error status depending on which of the two terminal hooks
get called.

We'll also emit a trace event for each resource instance we plan. Ideally
these would have their own spans too but we're currently limited by the
design of the terraform.Hook API which wasn't designed with tracing spans
in mind, and so we'll just use point-in-time events for now and rely on
manual time-based correlation to try to associate downstream spans with
particular resource instance operations if needed.