plans: ResourceInstanceChange.ObjectAddr
We've now many times made the mistake of thinking that the resource
instance address is a suitable unique identifier for a resource instance
change in a plan -- the name certainly seems to suggest it -- but really
these represent changes to resource instance _objects_ and so it's
important to also include the DeposedKey field as part of the unique
identifier for a change.

Now we'll expose ObjectAddr as a new method that returns the two together
for convenient use as a single address value, along with adding some
commentary to the fields to make it clearer that Addr is not sufficient.
Ideally we'd rename Addr to make it less of an attractive nuisance, but
that was too invasive a change for today so will need to wait for another
time.