stackeval: A FIXME about emitting prior state snapshots during apply
Currently we're not explicitly writing the results from the refresh phase
to the raw state or state description, which means that a refresh-only
plan wouldn't do anything and even in a normal plan there are some
edge-cases where we'll end up not saving the refreshed state.

This is just a note to remind us of that in later work.