configs: remove redundant provider reqs
This block of code iterated through import blocks, adding to reqs the
implied provider for each import target.
This is redundant, because for any import block, either there exists a
corresponding resource block in config, or there does not and config
will be generated.
If the resource block already exists, the provider requirement will be
added by code earlier in addProviderRequirements. If the resource
block does not exist, the provider requirement will be added by the code
block directly following this one.
The above applies whether or not the target of the import block is in a
child module. If a child module has a different local name for a
provider, that provider must be in required_providers, so again it will
be added to the provider requirements elsewhere in
addProviderRequirements.