rpcapi(Stacks): don't send empty moved/import messages
The "progress hooks" mechanism was unconditionally allocating messages
to represent the moved and imported cases and then conditionally populating
the fields inside.

That causes an admittedly-small friction on the client side because they
need to test whether the string fields inside are non-empty instead of
just checking whether the overall message field was populated at all.

To smooth that for clients, we'll instead leave these unset unless there's
actually something to report. For clients written in Go, that means that
the top-level fields will be nil when not relevant, which is easy to
test for and relatively idiomatic.