stacks+rpcapi: Wire through the "plan mode" to the stacks runtime
Previously we were accepting it at the API layer but then silently
discarding it and always creating a "normal" plan. Now we'll pass the
caller's selected planning mode all the way through to the modules runtime
in component plans.

This isn't actually quite sufficient to properly support destroy-mode plans
since the stacks runtime itself doesn't yet really understand how to
deal with things being destroyed, but this is intended to be just enough
for those working on callers of this API to see _some_ effect from setting
the planning mode, even though it's far from the full effect expected.

More work on destroying will follow in subsequent commits.