states: State.AllResourceInstanceObjectAddrs is back!
In the previous commit we renamed AllResourceInstaceObjectAddrs to
AllManagedResourceInstanceObjectAddrs to reflect that it filters out any
resource instances that aren't "managed".

This now adds back in a State.AllResourceInstanceObjectAddrs method which
does what the other one sounded like it did before: returns all of the
known addresses without any filtering whatsoever.