rpcapi: Remove Stacks.FindStackConfigurationProviders function
This was unimplemented because we didn't end up needing it. We might find
a use for something like it someday, but we'll wait and see what the
requirements are before we properly define it as part of the API.

(The current way we achieve "finding providers" is to use a dependency
lock file which is assumed to already contain exact provider selections.
The most likely reason for something like FindStackConfigurationProviders
to return is to allow building a tool for generating those dependency
lock files, which will require analyzing the configuration to find which
providers it actually depends on, but it remains to be seen exactly what
API would make sense for that operation.)