stackplan: Remove stale TODO about output value changes
This recorded that we needed to encode the type information for the
output values. In the end we didn't need to do anything special for that
because we're borrowing the pre-serialized data originally designed for
the traditional Terraform CLI saved plan format, and that encodes the
type information as part of the MessagePack encoding in-band.