stackeval: Remove stale todo about prior state in plans
This TODO was noting that an earlier version of this code was not
including any prior state information in the "modules runtime" plan. At
the time I wrote the comment I was expecting to handle that here as a
separate step, but I later refactored this to include the prior state as
part of the returned plan, since that's where the modules runtime expects
to find it during the apply phase anyway.

This comment is therefore now redundant, because no extra work is required
in this location to get the prior state populated.