stackstate: A helper for loading state during tests
The main entry point here assumes the caller is providing serialized state
objects wrapped in anypb.Any messages.

That's an inconvenient representation for hand-writing in tests, so the
new function LoadFromDirectProto allows skipping the deserialization steps
and instead has the caller write the wanted values as if they had already
been parsed/unmarshaled.