stackeval: Initial tests for type "Provider"
These tests focus on the behaviors related to deciding the set of dynamic
instances for a particular provider and producing the values that act as
references to providers.

These tests also revealed a small but important bug in ExprReferenceValue
in the for_each case, which is now fixed as part of this commit.