stackeval: inPromisingTask cancels context when test completes
This is just some extra insurance to reduce the risk of us leaving things
dangling after a test terminates. Although tests should ideally not rely
on this and should terminate background work properly as part of the test,
this should catch some things that might get left behind if a test ends
in an ungraceful way.