main: The "terraform rpcapi" command is experimental
For the moment this command is only used as part of the Terraform Stacks
workflow, which is in private preview. It's not yet suitable for any use
outside of that context because we're likely to make significant breaking
changes to this in response to feedback.

As usual with "experimental" features, the idea here is to allow this to
be included in alpha/dev builds while not yet including it in stable
releases or non-alpha prereleases. We're landing this here now primarily
so we can do remaining development in the public repository, now that the
conference announcement embargo for Stacks has been lifted.