read state lock info and lock holder from environment variables
The defult LockInfo object created by terraform duirng state lock does
not use the Info field present in StateLock object that is useful for
storing metadata. Neither does the it effectively set the Who field.
Currently the Who field default to user@host. Which isn't useful when
the lock is being created from inside a docker image or in a CI/CD
pipeline. In those cases, the user might find it userful if the Who
field was instead configurable.

This commit adds the feature of reading Info and Who fields for the
LockInfo object from the TF_LOCK_METADATA and TF_LOCK_OWNER_ID
environment variables if present.