core: NodeAbstraceResourceInstance.checkPreventDestroy returns diagnostics
This method was returning diagnostics wrapped inside an error, and all
of its callers were then immediately unwrapping it and using the
diagnostics inside anyway.

To make that clearer, we'll just return diagnostics directly.

(It's likely that this was originally written as a function that returned
an error before we even _had_ the concept of diagnostics, and over time
we've gradually migrated all of the callers to be ready to accept
diagnostics but neglected to actually update the function itself.)