core: Plan and apply can optionally return an expression scope
This is a similar idea to terraform.Context.Eval, but rather than using
the special eval-only walk (which just reads data from the latest state
snapshot) we can now produce an evaluation scope based on the result of
either a plan or an apply operation, evaluating against either the planned
state or the new state respectively.

This will allow (in future commits) some code reuse for the test harness
and also the option to use "terraform console" to evaluate expressions
against planned state rather than only against prior state. However, as
of this commit there are not yet any direct callers of these new methods
except for the two new test cases.